Add files via upload
The goal of this project is to develop a flow chart for software that tracks student attendance in cohort meetings, records their scores and performance, and gathers their feedback on the program. Additionally, the application will analyze this data to predict if a student is likely to complete the program or drop out.